version: allow to build on non-unix platforms
[ upstream commit a4123c84c91735b2e2325720ba384aac8e1d5b6b ]

This package is a transitive dependency in cilium-cli, but currently it
fails to build on non-unix platforms (e.g. windows) due to its use of
golang.org/x/sys/unix.

However, we want still to be able to build it for some these platforms,
see e.g. https://github.com/cilium/cilium-cli/issues/231. Fix the build
by moving all unix-specific functionality (i.e. the Uname call) to a
separate file, protected by build tags.
