+++ pkg/version/version_test.go
@@ -20,17 +20,13 @@ import (
 	"runtime"
 	"testing"
 
-	"github.com/cilium/cilium/pkg/versioncheck"
-
-	"github.com/blang/semver/v4"
 	. "gopkg.in/check.v1"
 )
 
 // Hook up gocheck into the "go test" runner.
 func Test(t *testing.T) { TestingT(t) }
 
-type VersionSuite struct {
-}
+type VersionSuite struct{}
 
 var _ = Suite(&VersionSuite{})
 
@@ -106,30 +102,4 @@ func (vs *VersionSuite) TestVersionArchMatchesGOARCH(c *C) {
 	// var ciliumVersion is not set in tests, thus Version does not contain the cilium version,
 	// just check that GOOS/GOARCH are reported correctly, see #13122.
 	c.Assert(Version, Matches, ".* "+runtime.GOOS+"/"+runtime.GOARCH)
-}
-
-func (vs *VersionSuite) TestParseKernelVersion(c *C) {
-	mustHaveVersion := func(v string) semver.Version {
-		ver, err := versioncheck.Version(v)
-		c.Assert(err, IsNil)
-		return ver
-	}
-
-	var flagtests = []struct {
-		in  string
-		out semver.Version
-	}{
-		{"4.10.0", mustHaveVersion("4.10.0")},
-		{"4.10", mustHaveVersion("4.10.0")},
-		{"4.12.0+", mustHaveVersion("4.12.0")},
-		{"4.12.8", mustHaveVersion("4.12.8")},
-		{"4.14.0-rc7+", mustHaveVersion("4.14.0")},
-		{"4.9.17-040917-generic", mustHaveVersion("4.9.17")},
-		{"4.9.generic", mustHaveVersion("4.9.0")},
-	}
-	for _, tt := range flagtests {
-		s, err := parseKernelVersion(tt.in)
-		c.Assert(err, IsNil)
-		c.Assert(tt.out.Equals(s), Equals, true)
-	}
 }
